Create PersistentUploadedFileHandler with manual file removal
There are two new classes in this file.
1. PersistentUploadedFile, which is a copy of TemporaryUploadedFile with
one small change: the delete flag on the NamedTemporaryFile is set to
False, so that the temporary file must be removed manually.

2. PersistentFileUploadHandler, which is a copy of
TemporaryFileUploadHandler, but it uses the new PersistentUploadedFile
instead of TemporaryUploadedFile

I created these classes for a very specific reason, that relates to my
using Dropzone.js to upload files. I could only get Dropzone to work by
making an Ajax request to send the files, separate from the transfer
form's POST submission. It so happened in my view functions that the
temporary files were being deleted after they were processed in the
Ajax request on the Django backend. I needed the files to stay on the
file system until the POST request came through for the form, hence the
delete=False flag. Obviously, this is a less safe way to create temp
files since they are not removed automatically, but it should be
possible in the future to periodically clean the temporary upload folder
to make sure there are no persistent storage memory leaks. At this
point, I am removing the temp files after the Bagger finishes creating a
bag with the files.