Start with 1 other identifiers form
Since I want to be able to add forms to the formset, I will need to
duplicate a form within the formset. Now, this won't be possible without
the markup already being there. I might consider having the one form
there, but hidden, but I haven't decided yet. All I know is that the
markup must be there.