Remove tag_files option from bagger
I read over the Archivematica documentation and the BagIt specification
and it makes more sense to just throw all of the metadata into
bag-info.txt, rather than write tag files. The main reason is so that
the data can be searched later in Archivematica. If it's in a random tag
file, Archivematica will not be able to parse it and won't even look at
the contents of the file.