Store filenames, not absolute paths
Since we are storing paths to files in the database, it makes sense that
we should be able to move the files around without breaking anything.
That being said, when we store absolute file paths in the database, we
absolutely will break things when we move those files around.

Now, the full paths to the bags and the reports are generated only when
the user requests them, and the names only are stored in the database.