Improve logging messages
There's still one issue with the messages: When a function is called in
a job, the logs are sent to the regular log file, when they should be
sent to the rqworker log file. I'm not sure why the rq log handler
doesn't catch messages from functions called within it, but I will
figure it out.

It might make sense to make the bagging functions into async-only jobs
since they really shouldn't be run on the main thread. That way I could
explicitly set the name for getLogger.