Emit post-migration signal to create missing perms
If starting with a fresh database, the permissions for each object
created by django.contrib.auth do not exist until after the migration.
The permission creation is triggered by a post-migration signal.

This caused issues because the archivist_permissions migration requires
that the permissions exist. The work-around is to emit the post
migration signal while the migration runs to create those permissions.