Add a view to check file extension
This view will be used as a first line of defense in the transfer form
to verify if a file can be uploaded. The file extension is also checked
in the uploadfiles view, but does not have as much granularity over
checking single files.

These changes currently break the transfer form.