Add redundant file extension check
If a malicious user manages to circumvent the initial file extension
check, it is checked again during upload and the endpoint will not allow
the file to be uploaded.

I check the extension initially so that there's no partially-submitted
set of files, but if someone bypasses the initial checks, a
partially-submitted set of files is likely to occur if any one file is
invalid. This happens if one or more files are uploaded succesfully
before an invalid one is attempted to be uploaded.