Rename rights_statement_note to rights_statement_value
I made it so that only the form is aware that there is an "Other" type
of rights. The rights_statement_value is an integral part of the CAAIS
standard, so I brought it back in the places it was removed. I set the
rights_statement_type to the other_rights_statement_type when "Other" is
chosen.

I also don't see the point in including the rights_statement_note,
because people can put whatever they want into the
rights_statement_value. The right_statement_note will now be NULL in the
bag-info.txt, and empty in the metadata report. I simply set the note to
an empty string in the clean() method of the RightsForm.