Rewrite admin.py
Many changes had to be made to support the new models. The most
important change is that I've added views to the model admins, which is
the proper way to be providing custom functionality rather than my
previous way of switching on POST arguments.

Using views, the model history no longer gets polluted with
non-change-related changes.