Update packages, upgrade Django 3.1.8 -> 3.2.11
Django 3.1 will be coming to the end of its life at the end of 2022, so
I upgraded to Django 3.2, which is a LTS release. Django 3.2 will be
supported until April 2024.

I would have upgraded to Django v4, but support for Python 3.6 is
removed in v4, which we still need support for.

Other notable upgrades:
bagit 1.7.0 -> 1.8.1
django-countries 6.1.3 -> 7.2.1
django-formtools 2.2 -> 2.3
django-rq 2.3.2 -> 2.5.1
gunicorn 20.0.4 -> 20.1.0
mysql-connector-python 8.0.22 -> 8.0.23
python-decouple 3.3 -> 3.5

Dev (non-application) related upgrades:
pylint 2.6.0 -> 2.12.2
Sphinx 3.2.1 -> 4.3.2
sphinx-rtd-theme 0.5.0 -> 1.0.0

The requirements.txt and requirements.dev.txt were not properly
separated prior to this commit. Now, the requirements.txt has the bare
minimum packages required to run the app. All the requirements.dev.txt
packages are only required for building the documentation and linting.