Re-implement file system Bag update code
The application administrator can choose whether the BagIt bags are
updated on the file system in tandem with the database Bag by setting
the ALLOW_BAG_CHANGES setting to true if they want these changes to
occur, or false if they do not.

The changes propagated to the BagIt bag in the case that
ALLOW_BAG_CHANGES is true occur when these events occur:
- A level of detail is added to a Submission
- The level of detail is changed for a Submission
- An accession identifier is added to a Submission
- The accession identifier is changed for a Submission
- An Appraisal is added for a Submission
- An Appraisal is removed from a Submission

If a Submission does not have a Bag, or ALLOW_BAG_CHANGES is false,
these updates make no changes to the submission's bag.

In every case that a change *is* made to a BagIt bag, an event is logged
in Section 5 (Event Information) of the bag's metadata.