Run debugpy for app on port 8009
debugpy is now set to run when DEBUG is True, and ENABLE_DEBUGPY is set
to "1". It is still not possible to debug asynchronous jobs from
jobs.py, it may be necessary to start a separate debugpy instance on an
adjacent port to provide debugging capabilities on the RQ jobs.