Generate Submission synchronously
The submission generates so quickly that there is no point in doing it
asynchronously. This way, we can ensure that the submission exists on
the user's profile page when they are redirected, and we know the status
of the error and can re-direct to an error page when the submission
fails.