Simplify by removing mode context processor and MODE variable
The mode context processor and MODE variable in .env are not required.
The WEBPACK_MODE is set directly in the compose file, which will load
that variable into the environment when webpack is run.

The mode variable is also not required in the templates, since the name
of the bundled asset will not change whether in development or
production. The one thing that *will* change is whether the assets are
minified, but that does not affect anything Django is doing.