Re-organize Containerfile for better caching
The ordering of commands in the Containerfile have been re-organized so
that caching of layers is taken better advantage of.

For example, the npm install command now only needs to run if
package.json, package-lock.json, or webpack.config.js change.

And the Python dependencies only need to be installed if pyproject.toml,
poetry.lock, or README.md changes.

Previously, both the NodeJS and Python dependencies were being re-
installed any time a code change was made, which is inefficient for
caching purposes.

These changes should make it quicker to re-build the image which is
useful for iterating over changes, especially in production.