Migrate to webpack (#248)
* Bundle CSS and JS using webpack

* Test with Python dependencies and entrypoint

* Name bundled CSS dynamically

* Remove includes from base template

* Ensure bundles static assets are stored and referenced from the same place

* Configure behaviour of webpack based on whether in production or development

* Add entry for transferform JS in webpack configuration

* Remove base static assets from pipeline

* Add entry for profile JS in webpack configuration, change name of folder from forms to transferform

* Remove profile js from pipeline

* Remove dropzone JS from pipeline, remove profile js include

* Add entry for submission detail CSS in webpack configuration

* Move building of static assets call to entrypoint

* Switch to container image that has both Node and Python installed

* Add entry for submission group JS in webpack configuration

* Add entry for caais metadata admin static assets in webpack configuration

* Add entry for recordtransfer admin uploadedfile JS in webpack configuration

* Use admin classes to add static assets instead of using templates

* Broken: remove django-pipeline as a dependency

* Build webpack assets in dist/, tell Django they're there

Django does not seem to be aware of static assets that it did not pick
up during collectstatic. This commit changes a few things:

- Webpack is run *before* collectstatic
- Webpack assets are sent to dist/, not static/
- When collectstatic is run, assets in dist/ are copied to the static
  directory
- Collecstatic is told not to copy un-built .js and .css files in
  recordtransfer

Django can now find the static assets that webpack built, and a 404
error is no longer thrown for those static assets.

* Simplify by removing mode context processor and MODE variable

The mode context processor and MODE variable in .env are not required.
The WEBPACK_MODE is set directly in the compose file, which will load
that variable into the environment when webpack is run.

The mode variable is also not required in the templates, since the name
of the bundled asset will not change whether in development or
production. The one thing that *will* change is whether the assets are
minified, but that does not affect anything Django is doing.

* Re-organize Containerfile for better caching

The ordering of commands in the Containerfile have been re-organized so
that caching of layers is taken better advantage of.

For example, the npm install command now only needs to run if
package.json, package-lock.json, or webpack.config.js change.

And the Python dependencies only need to be installed if pyproject.toml,
poetry.lock, or README.md changes.

Previously, both the NodeJS and Python dependencies were being re-
installed any time a code change was made, which is inefficient for
caching purposes.

These changes should make it quicker to re-build the image which is
useful for iterating over changes, especially in production.

* Remove redundant check of yuglify binary

* Broken: minified image generation using webpack and sharp

* Broken: minified image generation using webpack and imagemin

* Convert images to webp format using imagemin

* Convert images to webp using webpack and sharp

* Revert "Convert images to webp format using imagemin"

This reverts commit 2103b7c9632e05aa335e47ab538f7ebc703eb581.

* Revert "Broken: minified image generation using webpack and imagemin"

This reverts commit 5d917ae27f248e31494833d5e51066f97ba14272.

* Revert "Merge branch 'temp-247' into 247-migrate-to-webpack"

This reverts commit 072e1059e6c5ccd3e70150409f3292814e7d385b, reversing
changes made to 24a5c2fe021611fe08112aef1247369d41374a68.

* Revert "Revert "Broken: minified image generation using webpack and imagemin""

This reverts commit 24a5c2fe021611fe08112aef1247369d41374a68.

* Revert "Revert "Convert images to webp format using imagemin""

This reverts commit 7e63c5f2ab61203c5725a3110f010cd6beccf7a2.

* Revert "Convert images to webp format using imagemin"

This reverts commit 2103b7c9632e05aa335e47ab538f7ebc703eb581.

* Revert "Broken: minified image generation using webpack and imagemin"

This reverts commit 5d917ae27f248e31494833d5e51066f97ba14272.

* Revert "Broken: minified image generation using webpack and sharp"

This reverts commit 5fd76eea6779e85a2d175815894ed38c3fd858c9.

* Remove redundant MODE setting

---------

Co-authored-by: Daniel Lovegrove <d.lovegrove11@gmail.com>