Simplify formset setup by refactoring
The formset setup was overly complex - this has been changed to be more
declarative and now uses vanilla JS only.

The method of adding CSS classes was overly complex, as well. This has
been simplified so that there is only one class added to a form row when
that row is about to be deleted (the new class is called
imminent-removal).