Change view to accept GET requests to download uploaded files
This fixes an issue with not being able to view uploaded files in
production.

Instead of passing the actual media URL to the frontend, a path to the
uploaded_file view is returned. This was previously the
delete_uploaded_file view (now allowing GET and DELETE requests).

uploaded_file view is returned. This was previously the
delete_uploaded_file view (now allowing GET and DELETE requests).

In development (i.e., when DEBUG=True), the path to the media URL is
returned like it was before. The Django development server can handle
serving media files itself so no major changes were needed here.

In production, the path to the media URL is returned in the header
X-Accel-Redirect. This tells NGINX where to find the file, and that it's
OK to serve this file.

This works because requests to /media/ can *only* come from inside the
app. Trying to request a media file by URL directly in production will
always return a 404. In this case, the user makes a GET request to the
uploaded_file URL, and then an X-Accel-Redirect response is generated
*internally*.