Remove jQuery (#331)
* Remove jquery-input-mask, install imask

This swaps the jQuery-based input mask with a new dependency that does
not rely on jQuery.

No functionality has been changed in this commit!

* Use floating-ui instead of jQuery tooltip

The code quite a bit more complex but it is a lot more explicit about
how the tooltips are created now.

* Use fetch() instead of $.post(), only allow GET in view

Switches the $.post() call in the dropzone form to a GET request using
the fetch API. Also, changes the /transfer/checkfile/ view to a GET-only
endpoint, since data is not POST-ed using that view ever.

* Remove $ from hideMediaLink and navbar

* Re-write dropzoneForm.js

* Simplify formset setup by refactoring

The formset setup was overly complex - this has been changed to be more
declarative and now uses vanilla JS only.

The method of adding CSS classes was overly complex, as well. This has
been simplified so that there is only one class added to a form row when
that row is about to be deleted (the new class is called
imminent-removal).

* Move explanations to About page

This completely removes the need for three modals. The one remaining
modal is used to add a new group.

* Use vanilla CSS/JS modal for submission group form

* Remove unused MODAL_MODE

* Process error in fetch()

* WIP! Moving submission group modal to separate .js file

* Remove dropzone code

* WIP! Move code to setupSubmissionGroupForm()

* Restore all previous functionality

* Fix context not being passed to .js

* Handle when description is empty

* WIP! Move code to separate files

* Wrap uppy setup in function

* Revert changes to caais admin code

If caais is to be used as a separate module in the future, we can't
assume that someone will install IMask. Django's jQuery, on the other
hand, can be relied on to be included.

* Add air-datepicker dependency

* Simplify form by using a date range picker

* Create a single entrypoint into transfer form

* Remove last bit of jQuery from recordtransfer

* Remove jQuery CDN scripts

* Fix other province field

* Pass formset prefix to setupFormset()

* Fix other field on rights for initial forms only

* Fix other field on rights for new forms

* Fix empty other rights field allowed to submit

* Fix unsaved changes modal not appearing

* Fix formset on otheridentifiers form

* Fix h2 styles after merging master

* Add devtool for debugging in browser

* Shut off devtool on production

* Fix masking not working in CAAIS admin

* Fix misleading docstring

* Fix failing test

* Remove unused CSS

* Clear inputs and error messages on new form

* Include new .CSS files

* WIP: Trying to fix e2e tests

* Fix extra formset being added when resuming transfer

The NUM_EXTRA_FORMS was missing from the Javascript context - I added
it back and am deleting the last form if it's zero (the same way this
form is being handled on the master branch).

* Fix date widget entry when USE_DATE_WIDGETS=False

Only one date can be added for now.

* Disallow future dates, update error message

* Add 0.2s fade in-out transition to tooltip

* Simplify formset setup by using initial forms

* Add validation for date_of_materials

* Apply JS suggestions from code review

Co-authored-by: Kelvin Oo <kelvinoo123@gmail.com>

* Set input to readonly when using date widgets

* Auto-format file

---------

Co-authored-by: Kelvin Oo <kelvinoo123@gmail.com>