WIP: Get scheduling working
This adds another container to run the scheduler in parallel with RQ and
the app. The scheduler needs to run in a separate thread/process because
of the way it's implemented:

- It *polls* for new jobs in an infinite loop
- If a new job is found, it gets added to the queue

At that point, rq will pick up the job and run it.

Theoretically, this could be run in a separate thread in the rq or the
app container but that's an anti-pattern in containerized design.

Outstanding items:

- The log messages from schedule_cleanup are not being captured
- The expired upload session filtering needs to be updated
- register_scheduled_jobs should use the setting in django.conf instead
  of the hard-coded cron string
- The command should be given a more generic name since we may have more
  scheduled jobs than just the cleanup in the future
- An RQ scheduler needs to be added to the prod compose file