WIP: stop including conflicting modal styles
Tailwind already includes a .modal class, and there is a .modal class
defined in modal.css. I am no longer including the modal.css file so
that there are no conflicts between these styles, but this also means
that the modal in the submission form is broken.

We will want to replace the group creation modal in the submission form
with the same one on the profile page.