Simplify webpack config with main + vendor bundles (#607)
* Replace base bundle with main entrypoint

* Replace submission_form bundle with main entrypoint

* Replace submission_group bundle with main entrypoint

* Replace submission_detail entry point with main entrypoint

* Replace profile entry point with main entrypoint

* Revert "Replace profile entry point with main entrypoint"

This reverts commit 6ca19dca22193d164bbc8ee533f475ae732ade4b.

* Replace profile entry point with main entrypoint

* Move tailwind to new main.css file

* Simplify CSS by removing two unnecessary small files

There were two CSS files I found that had very few styles in them. These
files each contained one CSS class that could easily be replaced with a
class already provided by tailwindcss.

We have too many small CSS files so this should help start simplifying
the CSS of our app.

* Add back required submission_detail entrypoint

The submission_detail page is special in that it shares none of the
styling with the main application.

We'll want to address this in the future, but for now we still need a
separate entrypoint for this page.

* WIP: stop including conflicting modal styles

Tailwind already includes a .modal class, and there is a .modal class
defined in modal.css. I am no longer including the modal.css file so
that there are no conflicts between these styles, but this also means
that the modal in the submission form is broken.

We will want to replace the group creation modal in the submission form
with the same one on the profile page.

* Re-include the modal.css in index.js

* Remove the commented-out modal.css include in submission_form/index.js

* Rename the .modal class in modal.css to .recordtransfer-modal

* Rename modals being repeated in tailwind and change its instances

* Add css for modal-content required for daisyUI container

* Use splitChunks plugin so all node_modules have seperate vectors.bundle.js

* Remove splitchunks for vendors.bundle.js for now

* Add border for delect (Dropdowns)

* Split main entry point into main.bundle.js and vendor.bundle.js

* Renamed submission_group_form.js to index.js

* Remove extra css added to base.css

* Remove unecessary css imports in submission_form/index.js

* Remove extra imports for .css files when already imported by super

* Apply suggestions from code review

Co-authored-by: Daniel Lovegrove <d.lovegrove11@gmail.com>

* Rename functions to be imported in index.js

* Rename functions to be imported in index.js

* Assign windox.htmx to imported htmx

* Add styles to links to override tailwind/daisyUI styles

* Add daisyUI styling to the links of Submission sent page

* Make textarea and input have same font size

* Undo changes to base .css

* Fix djlint errors in html files changes

* Change colour of buttons and links on submission form

---------

Co-authored-by: Daniel Lovegrove <d.lovegrove11@gmail.com>