Override WEBPACK_LOADER in E2E tests
The unit tests use a fake webpack loader, but the E2E still need the
real loader.

This solution is incomplete right now because the path to
webpack-stats.json needs to be corrected.

And we can't use BASE_DIR right now in the tests because it's set to a
random temp folder.