Fix commented-out test
This test was not working because datetime.now() does not return times
consistent with timezone.now(). The django-axes code uses
timezone.now(), not datetime.now(), and changing the test to freeze time
based on timezone.now() has fixed the test.

I found this by debugging the axes_attempt_time property on the request
object. Here is a series of three invalid attempts:

datetime.datetime(2025, 7, 28, 16, 31, 32, 379869, tzinfo=datetime.timezone.utc)
datetime.datetime(2025, 7, 28, 16, 31, 58, 777520, tzinfo=datetime.timezone.utc)
datetime.datetime(2025, 7, 28, 16, 31, 58, 777520, tzinfo=datetime.timezone.utc)

And then the frozen time:

FakeDatetime(2025, 7, 28, 12, 33, 30, 707743, tzinfo=datetime.timezone.utc)

Note that the frozen time is four hours *before* the login attempts, so
this wasn't working (the test was expecting that the frozen time was 1
hour *after* the last login attempt).

One thing I wasn't expecting was that the response is a 302, instead of
a 200. Not sure if this is the expected behaviour or not.