Prevent brute force login attempts using `django-axes` (#856)
* Add django-axes dependency with ipware support

* Add Axes configuration and update authentication backends

* Make custom lockout view for failed login attempts

* Add docstring to lockout function

* Add tests for user lockout functionality after failed login attempts

* Disable axes by default in tests

* Enable AXES settings for login tests and comment out lockout period tests

* Add logging for account activation errors and specify backend in login

* Adjust test checking reset after successful login

* Fix commented-out test

This test was not working because datetime.now() does not return times
consistent with timezone.now(). The django-axes code uses
timezone.now(), not datetime.now(), and changing the test to freeze time
based on timezone.now() has fixed the test.

I found this by debugging the axes_attempt_time property on the request
object. Here is a series of three invalid attempts:

datetime.datetime(2025, 7, 28, 16, 31, 32, 379869, tzinfo=datetime.timezone.utc)
datetime.datetime(2025, 7, 28, 16, 31, 58, 777520, tzinfo=datetime.timezone.utc)
datetime.datetime(2025, 7, 28, 16, 31, 58, 777520, tzinfo=datetime.timezone.utc)

And then the frozen time:

FakeDatetime(2025, 7, 28, 12, 33, 30, 707743, tzinfo=datetime.timezone.utc)

Note that the frozen time is four hours *before* the login attempts, so
this wasn't working (the test was expecting that the frozen time was 1
hour *after* the last login attempt).

One thing I wasn't expecting was that the response is a 302, instead of
a 200. Not sure if this is the expected behaviour or not.

* Do not reset cool off period upon failed login during lockout, update tests

* Do not display lockout time to user

* Remove unused import of settings from account.py

* Increase lockout time and document main Axes settings

* Add Axes settings verification

* Show warning when close to lockout

* Add tests checking for warning and error trigger events

* Clarify failures being possibly None

* Document AXES_WARNING_THRESHOLD setting

* Apply suggestion to pluralize message

Co-authored-by: Daniel Lovegrove <d.lovegrove11@gmail.com>

* Add ngettext import

---------

Co-authored-by: Daniel Lovegrove <d.lovegrove11@gmail.com>