Update types in recordtransfer admin
Some refactoring was needed to remove all the errors. I created a new
mixin for uploaded file admins so that the fields: attribute had the
proper types.

I also refactored the save logic for resetting a SiteSetting. This is
because changeform_view needed a TemplateResponse, but an
HttpResponseRedirect was being returned. I made use of save_model and
response_change, which has the same behaviour as it did, but now
conforms to all the proper typings.