Upgrade python to 3.10, 3.11 (#919)
* Bump python version from 3.10 -> 3.11

We're now supporting 3.10 and 3.11, not 3.9 and 3.10.

* Update actions with new python version

* Update types in recordtransfer admin

Some refactoring was needed to remove all the errors. I created a new
mixin for uploaded file admins so that the fields: attribute had the
proper types.

I also refactored the save logic for resetting a SiteSetting. This is
because changeform_view needed a TemplateResponse, but an
HttpResponseRedirect was being returned. I made use of save_model and
response_change, which has the same behaviour as it did, but now
conforms to all the proper typings.

* Fix docstring

* Use gettext_lazy - see #921

* Remove unused import

* Add missing mixin

Co-authored-by: Kelvin Oo <kelvinoo123@gmail.com>

---------

Co-authored-by: Kelvin Oo <kelvinoo123@gmail.com>