Add tests for read-only uploaded file view
I struggled getting the tests to run with the proper view. The urls.py
seems to be loaded once whenever the tests run, and getting it to reload
was not working whatsoever. What did work was overriding the URL config
for the one test class where I needed it overridden.