Fix disabling uploads breaks file related admins (#1022)
* Add read-only uploaded file view

This view is loaded in when file uploads are disabled to still
allow files to be viewed. This is in case file uploads used to
be allowed, but are now not allowed.

* Add tests for read-only uploaded file view

I struggled getting the tests to run with the proper view. The urls.py
seems to be loaded once whenever the tests run, and getting it to reload
was not working whatsoever. What did work was overriding the URL config
for the one test class where I needed it overridden.

* Remove unnecessary wizard setup

* Remove duplicate setup