Fix build by running tailwind CLI directly
There are issues with the bun-tailwind-plugin that were causing the
styles to build incorrectly. I've updated the build script so that JS
files are built by Bun, and Tailwind builds the CSS files. These assets
are then listed in the same webpack-stats.json file.

I've removed the images because we don't really need to be building the
images like this.